+++ windows/redpanda-git-askpass/redpanda-git-askpass_private.h
@@ -1,0 +1,23 @@
+/* THIS FILE WILL BE OVERWRITTEN BY Red Panda C++ */
+/* DO NOT EDIT ! */
+
+#ifndef redpanda-git-askpass_private_rc
+#define redpanda-git-askpass_private_rc
+
+/* VERSION DEFINITIONS */
+#define VER_STRING	"1.0.0.0"
+#define VER_MAJOR	1
+#define VER_MINOR	0
+#define VER_RELEASE	0
+#define VER_BUILD	0
+#define COMPANY_NAME	""
+#define FILE_VERSION	""
+#define FILE_DESCRIPTION	"Developed using the Red Panda C++ IDE"
+#define INTERNAL_NAME	""
+#define LEGAL_COPYRIGHT	""
+#define LEGAL_TRADEMARKS	""
+#define ORIGINAL_FILENAME	""
+#define PRODUCT_NAME	""
+#define PRODUCT_VERSION	""
+
+#endif /*redpanda-git-askpass_private_rc*/