+++ windows/redpanda-git-askpass/resource.h
@@ -1,0 +1,7 @@
+#ifndef IDC_STATIC
+#define IDC_STATIC (-1)
+#endif
+
+#define IDD_MAIN_DIALOG                         101
+#define ID_TXT_PASSWORD                         40000
+#define ID_TXT_PROMPT                           40001