+++ NEWS.md
@@ -9,6 +9,7 @@ Red Panda C++ Version 0.14.4
   - enhancement: git - remotes
   - enhancement: rename "open folder" to "choose working folder"
   - enhancement: let user choose app theme when first run
+  - enhancement: git - pull / push / fetch
 
 Red Panda C++ Version 0.14.3
   - fix: wrong code completion font size, when screen dpi changed